refactor: search response now includes total + return types included for webSearch in the repo layer
